+++ cache/in-memory/src/event/channel.rs
@@ -203,37 +203,13 @@ impl UpdateCache for ChannelUpdate {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use twilight_model::{
-        channel::{ChannelType, TextChannel},
-        gateway::event::Event,
-        id::{ChannelId, GuildId},
-    };
-
-    fn guild_channel_text() -> (GuildId, ChannelId, GuildChannel) {
-        let guild_id = GuildId(1);
-        let channel_id = ChannelId(2);
-        let channel = GuildChannel::Text(TextChannel {
-            guild_id: Some(guild_id),
-            id: channel_id,
-            kind: ChannelType::GuildText,
-            last_message_id: None,
-            last_pin_timestamp: None,
-            name: "test".to_owned(),
-            nsfw: false,
-            parent_id: None,
-            permission_overwrites: Vec::new(),
-            position: 3,
-            rate_limit_per_user: None,
-            topic: None,
-        });
-
-        (guild_id, channel_id, channel)
-    }
+    use crate::test;
+    use twilight_model::gateway::event::Event;
 
     #[test]
     fn test_channel_delete_guild() {
         let cache = InMemoryCache::new();
-        let (guild_id, channel_id, channel) = guild_channel_text();
+        let (guild_id, channel_id, channel) = test::guild_channel_text();
 
         cache.cache_guild_channel(guild_id, channel.clone());
         assert_eq!(1, cache.0.channels_guild.len());
@@ -254,7 +230,7 @@ mod tests {
     #[test]
     fn test_channel_update_guild() {
         let cache = InMemoryCache::new();
-        let (guild_id, channel_id, channel) = guild_channel_text();
+        let (guild_id, channel_id, channel) = test::guild_channel_text();
 
         cache.update(&ChannelUpdate(Channel::Guild(channel)));
         assert_eq!(1, cache.0.channels_guild.len());
